Simplify math functions
'low' precision functions were accurate enough that there is no point in
overcomplicating it
