+++ apps/core/include/color/delta_e.h
@@ -4,7 +4,7 @@
 #include <vector>
 #include "color/lab.h"
 #include "color/simd_detect.h"
-#include "math.hpp"
+#include "simd_math.h"
 
 namespace delta {
